Fix #662: split AppDelegate into +PanelSetup and +StatusItem extensions
- Extract panel construction, PanelChromeView wiring, KVO observation,
  and Combine subscriptions into AppDelegate+PanelSetup.swift
- Extract status-item creation, icon updates, and menuBarImage helper
  into AppDelegate+StatusItem.swift
- Slim AppDelegate.swift down to property declarations, lifecycle entry
  point (applicationDidFinishLaunching calls the two setup helpers),
  OAuth URL callback, panel resize/navigate/key/dismiss/toggle/open logic
- All shared state remains `internal` (not private) so Swift can reach
  it across extension files — consistent with existing pattern in
  AppDelegate+Navigation.swift
