+++ Sources/RunnerBar/App/AppDelegate+PanelSetup.swift
@@ -1,0 +1,95 @@
+import AppKit
+import Combine
+import SwiftUI
+
+// MARK: - AppDelegate + Panel Setup
+//
+// Owns NSPanel construction, PanelChromeView wiring, KVO on
+// preferredContentSize, and Combine subscriptions that drive icon/store updates.
+// Called once from applicationDidFinishLaunching via setupPanel().
+//
+// ❌ NEVER inline this back into AppDelegate.swift.
+// ❌ NEVER call setupPanel() more than once.
+
+extension AppDelegate {
+
+    // MARK: Panel construction
+
+    /// Builds the NSPanel, embeds the SwiftUI hosting controller inside
+    /// PanelChromeView, wires KVO, and starts all Combine subscriptions.
+    func setupPanel() {
+        let controller = NSHostingController(rootView: mainView())
+        controller.sizingOptions = .preferredContentSize
+        controller.view.autoresizingMask = [.width, .height]
+        hostingController = controller
+
+        let initW = Self.initPanelWidth
+        let chromeView = PanelChromeView(
+            frame: NSRect(x: 0, y: 0, width: initW, height: 300 + arrowHeight)
+        )
+        chromeView.addSubview(controller.view)
+        chrome = chromeView
+
+        let newPanel = KeyablePanel(
+            contentRect: NSRect(x: 0, y: 0, width: initW, height: 300 + arrowHeight),
+            styleMask: [.borderless, .nonactivatingPanel],
+            backing: .buffered,
+            defer: false
+        )
+        newPanel.contentView = chromeView
+        newPanel.isOpaque = false
+        newPanel.backgroundColor = NSColor(white: 1, alpha: 0.001)
+        newPanel.hasShadow = true
+        newPanel.level = .popUpMenu
+        newPanel.collectionBehavior = [.canJoinAllSpaces, .fullScreenAuxiliary]
+        newPanel.animationBehavior = .none
+        panel = newPanel
+
+        setupKVO(controller: controller)
+        setupCombineSubscriptions()
+    }
+
+    // MARK: KVO
+
+    /// Observes `preferredContentSize` on the hosting controller so that any
+    /// SwiftUI layout change triggers a panel reposition without a visible jump.
+    private func setupKVO(controller: NSHostingController<AnyView>) {
+        sizeObservation = controller.observe(
+            \.preferredContentSize,
+            options: [.new]
+        ) { [weak self] _, change in
+            guard let size = change.newValue, size.height > 0 else { return }
+            DispatchQueue.main.async { self?.resizeAndRepositionPanel() }
+        }
+    }
+
+    // MARK: Combine subscriptions
+
+    /// Wires all Combine publishers that need to react during the app lifetime.
+    private func setupCombineSubscriptions() {
+        LocalRunnerStore.shared.$runners
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] _ in
+                guard let self else { return }
+                self.observable.reload(localRunnerStore: LocalRunnerStore.shared)
+            }
+            .store(in: &cancellables)
+
+        RunnerStore.shared.onChange = { [weak self] in
+            guard let self else { return }
+            log("AppDelegate › onChange fired — panelIsOpen=\(self.panelIsOpen) actions=\(RunnerStore.shared.actions.count) jobs=\(RunnerStore.shared.jobs.count)")
+            self.updateStatusIcon()
+            self.observable.reload(localRunnerStore: LocalRunnerStore.shared)
+        }
+        RunnerStore.shared.start()
+
+        ScopeStore.shared.didMutate
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] in
+                guard let self else { return }
+                log("AppDelegate › ScopeStore.didMutate — restarting RunnerStore")
+                RunnerStore.shared.restart()
+            }
+            .store(in: &cancellables)
+    }
+}
