+++ Sources/RunnerBar/App/AppDelegate+StatusItem.swift
@@ -1,0 +1,46 @@
+import AppKit
+
+// MARK: - AppDelegate + Status Item
+//
+// Owns NSStatusItem creation, menu-bar icon updates, and the menuBarImage
+// helper that maps AggregateStatus to the correct SF Symbol.
+// Called once from applicationDidFinishLaunching via setupStatusItem().
+//
+// ❌ NEVER inline this back into AppDelegate.swift.
+// ❌ NEVER call setupStatusItem() more than once.
+
+extension AppDelegate {
+
+    // MARK: Status item setup
+
+    /// Creates the NSStatusItem, sets the initial icon, and wires the toggle action.
+    func setupStatusItem() {
+        statusItem = NSStatusBar.system.statusItem(withLength: NSStatusItem.squareLength)
+        if let button = statusItem?.button {
+            button.image = menuBarImage(for: .allOffline)
+            button.action = #selector(togglePanel)
+            button.target = self
+        }
+    }
+
+    // MARK: Icon updates
+
+    /// Updates the menu-bar icon to reflect the current aggregate runner status.
+    /// ❌ NEVER filter by !isDimmed only — dimmed groups can still have in-progress jobs.
+    /// ❌ NEVER read RunnerStore.shared.jobs here — it is almost always empty.
+    /// ❌ NEVER call makeStatusIcon() — it no longer exists; use menuBarImage(for:).
+    /// If you are an agent or human, DO NOT REMOVE THIS COMMENT, YOU ARE NOT ALLOWED
+    /// UNDER ANY CIRCUMSTANCE.
+    func updateStatusIcon() {
+        statusItem?.button?.image = menuBarImage(for: RunnerStore.shared.aggregateStatus)
+    }
+
+    // MARK: Image helper
+
+    /// Returns the SF Symbol image for the given aggregate status.
+    func menuBarImage(for status: AggregateStatus) -> NSImage {
+        NSImage(systemSymbolName: status.symbolName, accessibilityDescription: nil)
+            ?? NSImage(systemSymbolName: "circle", accessibilityDescription: nil)
+            ?? NSImage()
+    }
+}
